After refactoring ParticleSystem

+++ PXCore/Particle/ParticleSystemBase.cpp
@@ -8,7 +8,7 @@ namespace Core::Particle {
 		_image.create(_settings.size.x, _settings.size.y, _TRANSPARENT);
 		_texture.loadFromImage(_image);
 		_sprite = sf::Sprite(_texture);
-		_sprite.setOrigin(_settings.size.x / 2.0f, _settings.size.y);
+		_sprite.setOrigin(static_cast<float>(_settings.size.x) / 2.0f, static_cast<float>(_settings.size.y));
 		_sprite.setPosition(_settings.position);
 	}
 	void ParticleSystemBase::AddParticles(unsigned int particles) {
@@ -27,12 +27,6 @@ namespace Core::Particle {
 		Clear();
 		PrepareTexture();
 		window.draw(_sprite);
-//#ifdef DEBUG
-//		sf::CircleShape circle(1.0);
-//		circle.setPosition(_settings.position);
-//		circle.setFillColor(sf::Color::Black);
-//		window.draw(circle);
-//#endif
 	}
 	bool ParticleSystemBase::ToDelete() const {
 		return _particles.size() == 0;
@@ -40,16 +34,6 @@ namespace Core::Particle {
 	void ParticleSystemBase::PrepareTexture() {
 		for (const auto& particle : _particles)
 			particle->Draw(_image);
-//#ifdef DEBUG
-//		for (int x = 0; x < _settings.size.x; x++) {
-//			_image.setPixel(x, 0, sf::Color::Black);
-//			_image.setPixel(x, _settings.size.y - 1, sf::Color::Black);
-//		}
-//		for (int y = 0; y < _settings.size.y; y++) {
-//			_image.setPixel(0, y, sf::Color::Black);
-//			_image.setPixel(_settings.size.x - 2, y, sf::Color::Black);
-//		}
-//#endif
 		_texture.update(_image);
 	}
 	void ParticleSystemBase::Clear() {
